优化了认证投资人审核列表的接口

+++ src/main/java/com/lhjl/tzzs/proxy/mapper/InvestorsApprovalMapper.java
@@ -34,5 +34,5 @@ public interface InvestorsApprovalMapper extends OwnerMapper<InvestorsApproval> 
 	 * @param body
 	 * @return
 	 */
-	Integer findApprovalListCount(InvestorsApprovalInputDto body);
+	Long findApprovalListCount(InvestorsApprovalInputDto body);
 }
